+++ src/TranslationLoaders/Db.php
@@ -16,7 +16,7 @@ class Db implements TranslationLoader
 
     protected function getConfiguredModelClass(): string
     {
-        $modelClass = config('laravel-translation-loader.model');
+        $modelClass = config('translation-loader.model');
 
         if (! is_a(new $modelClass, LanguageLine::class)) {
             throw InvalidConfiguration::invalidModel($modelClass);