+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `laravel-translation-loader` will be documented in this f
 ## 2.0.0 - 2017-XX-XX
 
 - add support for Laravel 5.5, dropped support for Laravel 5.4
+- rename config file from `laravel-translation-loader` to `translation-loader`
 
 ## 1.2.0 - 2017-07-07